refactor(main): extract the authenticated fetcher into engine-fetcher.ts
Move-and-parameterize of the 225-line anonymous fetcher closure (a third
of create-app-engine.ts) into apps/main/lib/engine-fetcher.ts:
createEngineFetcher({auth, clockSkew, emitTransport, fetch?, now?}).
Deliberately app-local with no generalization — no auth-transport
abstraction, no error classes, no retry dials. The mutate-in-place
semantics of the auth options and the shared clock-skew generation gate
are preserved by reference; the 401/refresh arc, level rubric, and
census total fetch move verbatim with their load-bearing comments. The
16-line declaration-ordering comment dies with the extraction (the
dependency is now a parameter).

38 fetcher-behavior tests move to engine-fetcher.test.ts and run against
the factory directly — no jest.resetModules, no sync-engine/database
mocks, no mock-call reach-ins. The two cache-and-fetcher wiring tests
stay in create-app-engine.test.ts.

Validation: apps/main targeted jest 59/59 (21 host + 38 fetcher), root
lint clean. Net production change +24 lines; total diff is dominated by
the moved test lines (budget overage reviewed and authorized).

Part of the C5-C8 architecture investigation (2026-08-05).
