+++ apps/main/lib/create-app-engine.test.ts
@@ -110,272 +110,6 @@ function loadCreateAppEngine(
 }
 
 describe('createAppSyncEngine scope cache', () => {
-	it.each([
-		['orders', 'wc/v3/orders'],
-		['products', 'wc/v3/products'],
-		['customers', 'wcpos/v2/customers'],
-		['taxRates', 'wcpos/v2/taxes'],
-		['categories', 'wc/v3/products/categories'],
-		['brands', 'wc/v3/products/brands'],
-		['tags', 'wc/v3/products/tags'],
-		['coupons', 'wc/v3/coupons'],
-	])('fetches the %s census through its configured route', async (collection, route) => {
-		const fetch = jest.spyOn(globalThis, 'fetch').mockResolvedValue(
-			new Response(null, {
-				status: 200,
-				headers: { 'X-WP-Total': '17', 'content-length': '0' },
-			})
-		);
-		const { createAppSyncEngine, createRxdbSyncEngine, recordTransport } = loadCreateAppEngine();
-		createAppSyncEngine(BASE_OPTIONS);
-		const queryTotal = createRxdbSyncEngine.mock.calls[0]?.[0].queryTotal;
-
-		const total = await queryTotal?.fetchWooQueryTotal({
-			request: {
-				queryKey: `census:${collection}`,
-				endpoint: collection,
-				params: { ignored: 'value', page: 9, per_page: 50 },
-				totalHeader: 'X-WP-Total',
-			},
-		});
-
-		expect(total).toBe(17);
-		expect(fetch).toHaveBeenCalledWith(
-			`https://store.example.test/wp-json/${route}?ignored=value&page=1&per_page=1`,
-			expect.objectContaining({ headers: expect.any(Headers) })
-		);
-		expect(recordTransport).toHaveBeenCalledTimes(1);
-		fetch.mockRestore();
-	});
-
-	it('reports variations census as unsupported without making a request', async () => {
-		const fetch = jest.spyOn(globalThis, 'fetch');
-		const { createAppSyncEngine, createRxdbSyncEngine } = loadCreateAppEngine();
-		createAppSyncEngine(BASE_OPTIONS);
-		const queryTotal = createRxdbSyncEngine.mock.calls[0]?.[0].queryTotal;
-
-		const total = await queryTotal?.fetchWooQueryTotal({
-			request: {
-				queryKey: 'census:variations',
-				endpoint: 'variations',
-				params: { page: 1, per_page: 1 },
-				totalHeader: 'X-WP-Total',
-			},
-		});
-
-		expect(total).toBeNull();
-		expect(fetch).not.toHaveBeenCalled();
-		fetch.mockRestore();
-	});
-
-	it('reports an unknown census collection as unsupported without making a request', async () => {
-		const fetch = jest.spyOn(globalThis, 'fetch').mockResolvedValue(
-			new Response(null, {
-				status: 200,
-				headers: { 'X-WP-Total': '17', 'content-length': '0' },
-			})
-		);
-		const { createAppSyncEngine, createRxdbSyncEngine } = loadCreateAppEngine();
-		createAppSyncEngine(BASE_OPTIONS);
-		const queryTotal = createRxdbSyncEngine.mock.calls[0]?.[0].queryTotal;
-
-		const total = await queryTotal?.fetchWooQueryTotal({
-			request: {
-				queryKey: 'census:unknown',
-				endpoint: 'unknown',
-				params: { page: 1, per_page: 1 },
-				totalHeader: 'X-WP-Total',
-			},
-		});
-
-		expect(total).toBeNull();
-		expect(fetch).not.toHaveBeenCalled();
-		fetch.mockRestore();
-	});
-
-	it.each([null, '', '3.5', '-1', 'not-a-number'])(
-		'rejects an invalid X-WP-Total value %p',
-		async (header) => {
-			const headers = new Headers();
-			if (header !== null) headers.set('X-WP-Total', header);
-			const fetch = jest
-				.spyOn(globalThis, 'fetch')
-				.mockResolvedValue(new Response(null, { status: 200, headers }));
-			const { createAppSyncEngine, createRxdbSyncEngine } = loadCreateAppEngine();
-			createAppSyncEngine(BASE_OPTIONS);
-			const queryTotal = createRxdbSyncEngine.mock.calls[0]?.[0].queryTotal;
-
-			await expect(
-				queryTotal?.fetchWooQueryTotal({
-					request: {
-						queryKey: 'census:orders',
-						endpoint: 'orders',
-						params: {},
-						totalHeader: 'X-WP-Total',
-					},
-				})
-			).rejects.toThrow('Invalid X-WP-Total');
-			fetch.mockRestore();
-		}
-	);
-
-	it('wires diagnostics and records response metrics without reading the body', async () => {
-		const now = jest.spyOn(Date, 'now').mockReturnValueOnce(1_000).mockReturnValueOnce(1_025);
-		const response = new Response('do not read', {
-			status: 200,
-			headers: {
-				'content-length': '42',
-				'X-Server-Load': '[0.5,0.3,0.2]',
-			},
-		});
-		const fetch = jest.spyOn(globalThis, 'fetch').mockResolvedValue(response);
-		const {
-			createAppSyncEngine,
-			createRxdbSyncEngine,
-			appMetricsObserver,
-			recordTransport,
-			recordServerLoad,
-		} = loadCreateAppEngine();
-		createAppSyncEngine(BASE_OPTIONS);
-		const ports = createRxdbSyncEngine.mock.calls[0]?.[0];
-
-		const result = await ports?.fetcher?.('https://store.example.test/wp-json/wcpos/v2/products');
-
-		expect(result).toBe(response);
-		expect(ports?.diagnostics).toEqual(expect.any(Function));
-		expect(appMetricsObserver).toHaveBeenCalledWith({
-			type: 'transport.request',
-			level: 'info',
-			at: 1_025,
-			fields: {
-				durationMs: 25,
-				bytes: 42,
-				status: 200,
-				method: 'GET',
-				path: '/wp-json/wcpos/v2/products',
-			},
-		});
-		expect(recordTransport).toHaveBeenCalledWith({
-			atMs: 1_025,
-			durationMs: 25,
-			bytes: 42,
-			ok: true,
-			epoch: 0,
-		});
-		expect(recordServerLoad).toHaveBeenCalledWith(0.5, 0);
-		expect(response.bodyUsed).toBe(false);
-		now.mockRestore();
-		fetch.mockRestore();
-	});
-
-	it('logs server clock skew at most once per scope', async () => {
-		const now = jest.spyOn(Date, 'now').mockReturnValue(0);
-		const fetch = jest.spyOn(globalThis, 'fetch').mockResolvedValue(
-			new Response(null, {
-				status: 200,
-				headers: { Date: 'Thu, 01 Jan 1970 00:02:00 GMT' },
-			})
-		);
-		const { createAppSyncEngine, createRxdbSyncEngine, networkWarn } = loadCreateAppEngine();
-		createAppSyncEngine(BASE_OPTIONS);
-		const fetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-
-		await fetcher?.('https://store.example.test/wp-json/wcpos/v2/products');
-		await fetcher?.('https://store.example.test/wp-json/wcpos/v2/orders');
-
-		expect(networkWarn).toHaveBeenCalledTimes(1);
-		expect(networkWarn).toHaveBeenCalledWith('Server clock is 120s ahead of the device clock', {
-			context: {
-				skewSeconds: 120,
-				serverDate: '1970-01-01T00:02:00.000Z',
-				deviceDate: '1970-01-01T00:00:00.000Z',
-			},
-		});
-		now.mockRestore();
-		fetch.mockRestore();
-	});
-
-	it('does not let a late prior-site response suppress the new site warning', async () => {
-		const now = jest.spyOn(Date, 'now').mockReturnValue(0);
-		let resolvePriorFetch!: (response: Response) => void;
-		const priorFetch = new Promise<Response>((resolve) => {
-			resolvePriorFetch = resolve;
-		});
-		const fetch = jest
-			.spyOn(globalThis, 'fetch')
-			.mockReturnValueOnce(priorFetch)
-			.mockResolvedValueOnce(
-				new Response(null, {
-					status: 200,
-					headers: { Date: 'Thu, 01 Jan 1970 00:03:00 GMT' },
-				})
-			);
-		const { createAppSyncEngine, createRxdbSyncEngine, networkWarn } = loadCreateAppEngine();
-		createAppSyncEngine(BASE_OPTIONS);
-		const priorFetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-		const priorRequest = priorFetcher?.('https://store.example.test/wp-json/wcpos/v2/products');
-
-		createAppSyncEngine(OTHER_SITE_OPTIONS);
-		const currentFetcher = createRxdbSyncEngine.mock.calls[1]?.[0].fetcher;
-		resolvePriorFetch(
-			new Response(null, {
-				status: 200,
-				headers: { Date: 'Thu, 01 Jan 1970 00:02:00 GMT' },
-			})
-		);
-		await priorRequest;
-		await currentFetcher?.('https://other.example.test/wp-json/wcpos/v2/products');
-
-		expect(networkWarn).toHaveBeenCalledWith(
-			'Server clock is 180s ahead of the device clock',
-			expect.any(Object)
-		);
-		now.mockRestore();
-		fetch.mockRestore();
-	});
-
-	it('does not let a late response from an earlier activation of the same scope suppress it', async () => {
-		const now = jest.spyOn(Date, 'now').mockReturnValue(0);
-		let resolvePriorFetch!: (response: Response) => void;
-		const priorFetch = new Promise<Response>((resolve) => {
-			resolvePriorFetch = resolve;
-		});
-		const fetch = jest
-			.spyOn(globalThis, 'fetch')
-			.mockReturnValueOnce(priorFetch)
-			.mockResolvedValueOnce(
-				new Response(null, {
-					status: 200,
-					headers: { Date: 'Thu, 01 Jan 1970 00:03:00 GMT' },
-				})
-			);
-		const { createAppSyncEngine, createRxdbSyncEngine, networkWarn } = loadCreateAppEngine();
-		createAppSyncEngine(BASE_OPTIONS);
-		const fetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-		const priorRequest = fetcher?.('https://store.example.test/wp-json/wcpos/v2/products');
-
-		createAppSyncEngine({
-			...BASE_OPTIONS,
-			scope: { ...BASE_OPTIONS.scope, storeId: 'store-2' },
-		});
-		createAppSyncEngine(BASE_OPTIONS);
-		resolvePriorFetch(
-			new Response(null, {
-				status: 200,
-				headers: { Date: 'Thu, 01 Jan 1970 00:02:00 GMT' },
-			})
-		);
-		await priorRequest;
-		await fetcher?.('https://store.example.test/wp-json/wcpos/v2/orders');
-
-		expect(networkWarn).toHaveBeenCalledWith(
-			'Server clock is 180s ahead of the device clock',
-			expect.any(Object)
-		);
-		now.mockRestore();
-		fetch.mockRestore();
-	});
-
 	it('does not let an auth retry adopt a later scope activation', async () => {
 		const now = jest.spyOn(Date, 'now').mockReturnValue(0);
 		let resolveRefresh!: (token: string) => void;
@@ -421,45 +155,6 @@ describe('createAppSyncEngine scope cache', () => {
 			'Server clock is 180s ahead of the device clock',
 			expect.any(Object)
 		);
-		now.mockRestore();
-		fetch.mockRestore();
-	});
-
-	it('classifies a conditional-GET 304 as a successful transport outcome', async () => {
-		const now = jest.spyOn(Date, 'now').mockReturnValueOnce(2_000).mockReturnValueOnce(2_010);
-		const response = new Response(null, {
-			status: 304,
-			headers: { etag: '"20:aa"' },
-		});
-		const fetch = jest.spyOn(globalThis, 'fetch').mockResolvedValue(response);
-		const { createAppSyncEngine, createRxdbSyncEngine, appMetricsObserver, recordTransport } =
-			loadCreateAppEngine();
-		createAppSyncEngine(BASE_OPTIONS);
-		const ports = createRxdbSyncEngine.mock.calls[0]?.[0];
-
-		await ports?.fetcher?.('https://store.example.test/wp-json/wcpos/v2/changes/sequence-log');
-
-		// Response.ok is false for 304, but an idle conditional poll answering
-		// Not Modified every tick must not be counted or logged as a failure.
-		expect(appMetricsObserver).toHaveBeenCalledWith({
-			type: 'transport.request',
-			level: 'info',
-			at: 2_010,
-			fields: {
-				durationMs: 10,
-				bytes: 0,
-				status: 304,
-				method: 'GET',
-				path: '/wp-json/wcpos/v2/changes/sequence-log',
-			},
-		});
-		expect(recordTransport).toHaveBeenCalledWith({
-			atMs: 2_010,
-			durationMs: 10,
-			bytes: 0,
-			ok: true,
-			epoch: 0,
-		});
 		now.mockRestore();
 		fetch.mockRestore();
 	});
@@ -576,160 +271,6 @@ describe('createAppSyncEngine scope cache', () => {
 	// observer exception propagated out of fetcher() and looked to the caller like
 	// a failed HTTP request — turning a request that actually succeeded into a
 	// failure (and, on the push path, a lost order).
-	it('returns the response even when a telemetry sink throws', async () => {
-		const fetch = jest
-			.spyOn(globalThis, 'fetch')
-			.mockResolvedValue(new Response(null, { status: 200 }));
-		const { createAppSyncEngine, createRxdbSyncEngine, appMetricsObserver } = loadCreateAppEngine();
-		appMetricsObserver.mockImplementation(() => {
-			throw new TypeError('observer exploded');
-		});
-		createAppSyncEngine(BASE_OPTIONS);
-		const fetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-
-		const response = await fetcher?.('https://store.example.test/wp-json/wcpos/v2/products');
-
-		expect(response?.status).toBe(200);
-		expect(appMetricsObserver).toHaveBeenCalled();
-		fetch.mockRestore();
-	});
-
-	it('records a network error and rethrows it', async () => {
-		const now = jest.spyOn(Date, 'now').mockReturnValueOnce(2_000).mockReturnValueOnce(2_040);
-		const networkError = new Error(
-			'network down for https://store.example.test/wp-json/wcpos/v2/products?authorization=secret'
-		);
-		const fetch = jest.spyOn(globalThis, 'fetch').mockRejectedValue(networkError);
-		const {
-			createAppSyncEngine,
-			createRxdbSyncEngine,
-			appMetricsObserver,
-			recordTransport,
-			networkWarn,
-		} = loadCreateAppEngine();
-		createAppSyncEngine(BASE_OPTIONS);
-		const fetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-
-		await expect(fetcher?.('https://store.example.test/wp-json/wcpos/v2/products')).rejects.toBe(
-			networkError
-		);
-
-		expect(appMetricsObserver).toHaveBeenCalledWith({
-			type: 'transport.request',
-			level: 'warn',
-			fields: {
-				durationMs: 40,
-				bytes: 0,
-				status: 0,
-				method: 'GET',
-				path: '/wp-json/wcpos/v2/products',
-			},
-		});
-		expect(recordTransport).toHaveBeenCalledWith({
-			atMs: 2_040,
-			durationMs: 40,
-			bytes: 0,
-			ok: false,
-			epoch: 0,
-		});
-		// The legacy networkLogger row is gone: the transport event now flows through
-		// the guarded log observer, so the failure lands as a wide terminal row with
-		// an outcome instead of a bare context blob (one writer per semantic event).
-		expect(networkWarn).toHaveBeenCalledWith('transport.request', {
-			context: expect.objectContaining({
-				type: 'transport.request',
-				method: 'GET',
-				path: '/wp-json/wcpos/v2/products',
-				status: 0,
-			}),
-			terminal: expect.objectContaining({ operationType: 'sync.http', outcome: 'failed' }),
-		});
-		now.mockRestore();
-		fetch.mockRestore();
-	});
-
-	it('does not persist expected sync aborts as errors', async () => {
-		const abort = Object.assign(new Error('aborted'), { name: 'AbortError' });
-		const fetch = jest.spyOn(globalThis, 'fetch').mockRejectedValue(abort);
-		const {
-			createAppSyncEngine,
-			createRxdbSyncEngine,
-			appMetricsObserver,
-			recordTransport,
-			networkWarn,
-		} = loadCreateAppEngine();
-		createAppSyncEngine(BASE_OPTIONS);
-
-		await expect(
-			createRxdbSyncEngine.mock.calls[0]?.[0].fetcher?.(
-				'https://store.example.test/wp-json/wcpos/v2/products'
-			)
-		).rejects.toBe(abort);
-		expect(appMetricsObserver).toHaveBeenCalledWith(
-			expect.objectContaining({
-				type: 'transport.request',
-				level: 'warn',
-				fields: expect.objectContaining({ status: 0 }),
-			})
-		);
-		expect(recordTransport).toHaveBeenCalledWith(expect.objectContaining({ ok: false }));
-		expect(networkWarn).not.toHaveBeenCalled();
-		fetch.mockRestore();
-	});
-
-	it('does not persist a row for a successful request, and never logs query credentials', async () => {
-		const fetch = jest
-			.spyOn(globalThis, 'fetch')
-			.mockResolvedValue(new Response(null, { status: 200 }));
-		const { createAppSyncEngine, createRxdbSyncEngine, networkInfo, appMetricsObserver } =
-			loadCreateAppEngine();
-		createAppSyncEngine({ ...BASE_OPTIONS, useJwtAsParam: true });
-		const fetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-
-		await fetcher?.('https://store.example.test/wp-json/wcpos/v2/push/orders?cursor=secret', {
-			method: 'POST',
-		});
-
-		// A succeeded HTTP attempt is deliberately NOT a durable row: the engine
-		// issues them continuously and they would evict the rest of the log inside
-		// the retention cap. Successful record pushes are covered by push.outcome.
-		expect(networkInfo).not.toHaveBeenCalled();
-		// It still reaches the metrics path, and only ever as a bare pathname —
-		// the query string carries the credential.
-		expect(appMetricsObserver).toHaveBeenCalledWith(
-			expect.objectContaining({
-				type: 'transport.request',
-				fields: expect.objectContaining({
-					method: 'POST',
-					path: '/wp-json/wcpos/v2/push/orders',
-					status: 200,
-				}),
-			})
-		);
-		expect(JSON.stringify(appMetricsObserver.mock.calls)).not.toContain('secret');
-		fetch.mockRestore();
-	});
-
-	it.each(['not-json', '{"load":0.5}', '["0.5"]'])(
-		'ignores malformed server load %s',
-		async (load) => {
-			const fetch = jest.spyOn(globalThis, 'fetch').mockResolvedValue(
-				new Response(null, {
-					status: 200,
-					headers: { 'X-Server-Load': load },
-				})
-			);
-			const { createAppSyncEngine, createRxdbSyncEngine, recordServerLoad } = loadCreateAppEngine();
-			createAppSyncEngine(BASE_OPTIONS);
-			const fetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-
-			await fetcher?.('https://store.example.test/wp-json/wcpos/v2/products');
-
-			expect(recordServerLoad).not.toHaveBeenCalled();
-			fetch.mockRestore();
-		}
-	);
-
 	it('returns the identical engine when the same scope is constructed twice', () => {
 		const { createAppSyncEngine, createRxdbSyncEngine } = loadCreateAppEngine();
 
@@ -1050,319 +591,6 @@ describe('createAppSyncEngine scope cache', () => {
 		const headers = init.headers as Headers;
 		expect(headers.get('Authorization')).toBeNull();
 		expect(headers.get('X-WCPOS')).toBe('1');
-		fetch.mockRestore();
-	});
-
-	it('refreshes after a 401 and retries once with the latest access token', async () => {
-		let accessToken = 'expired-token';
-		const credentials = {
-			getLatest: jest.fn(() => ({ access_token: accessToken })),
-		};
-		const refreshAuth = jest.fn(async () => {
-			accessToken = 'refreshed-token';
-			return accessToken;
-		});
-		const fetch = jest
-			.spyOn(globalThis, 'fetch')
-			.mockResolvedValueOnce(new Response(null, { status: 401 }))
-			.mockResolvedValueOnce(new Response(null, { status: 200 }));
-		const { createAppSyncEngine, createRxdbSyncEngine, appMetricsObserver, networkError } =
-			loadCreateAppEngine();
-		createAppSyncEngine({ ...BASE_OPTIONS, credentials, refreshAuth });
-		const fetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-
-		const response = await fetcher?.('https://store.example.test/wp-json/wcpos/v2/products');
-
-		expect(response?.status).toBe(200);
-		expect(refreshAuth).toHaveBeenCalledTimes(1);
-		// The arc id minted for the absorbed 401 is handed to the refresh layer so
-		// its "Session renewed" breadcrumb chains to these rows (#899).
-		expect(refreshAuth).toHaveBeenCalledWith({ operationId: expect.any(String) });
-		expect(fetch).toHaveBeenCalledTimes(2);
-		expect(networkError).not.toHaveBeenCalled();
-		// #899: the level reflects the SETTLED outcome. A 401 the refresh absorbed
-		// is forensic (debug + outcome 'recovered'), not a warn — and both attempts
-		// share the arc's operationId.
-		const arcCalls = refreshAuth.mock.calls as unknown as { operationId?: string }[][];
-		const arcId = arcCalls[0]?.[0]?.operationId;
-		expect(appMetricsObserver).toHaveBeenNthCalledWith(
-			1,
-			expect.objectContaining({
-				type: 'transport.request',
-				level: 'debug',
-				fields: expect.objectContaining({
-					status: 401,
-					outcome: 'recovered',
-					operationId: arcId,
-				}),
-			})
-		);
-		expect(appMetricsObserver).toHaveBeenNthCalledWith(
-			2,
-			expect.objectContaining({
-				type: 'transport.request',
-				level: 'debug',
-				fields: expect.objectContaining({ status: 200, operationId: arcId }),
-			})
-		);
-		const firstHeaders = fetch.mock.calls[0]?.[1]?.headers as Headers;
-		const retryHeaders = fetch.mock.calls[1]?.[1]?.headers as Headers;
-		expect(firstHeaders.get('Authorization')).toBe('Bearer expired-token');
-		expect(retryHeaders.get('Authorization')).toBe('Bearer refreshed-token');
-		fetch.mockRestore();
-	});
-
-	it('settles a 401 that persists after refresh as an error row (#899 exhaustion)', async () => {
-		let accessToken = 'expired-token';
-		const credentials = { getLatest: jest.fn(() => ({ access_token: accessToken })) };
-		const refreshAuth = jest.fn(async () => {
-			accessToken = 'refreshed-token';
-			return accessToken;
-		});
-		const fetch = jest
-			.spyOn(globalThis, 'fetch')
-			.mockResolvedValueOnce(new Response(null, { status: 401 }))
-			.mockResolvedValueOnce(new Response(null, { status: 401 }));
-		const { createAppSyncEngine, createRxdbSyncEngine, appMetricsObserver } = loadCreateAppEngine();
-		createAppSyncEngine({ ...BASE_OPTIONS, credentials, refreshAuth });
-		const fetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-
-		const response = await fetcher?.('https://store.example.test/wp-json/wcpos/v2/products');
-
-		expect(response?.status).toBe(401);
-		expect(refreshAuth).toHaveBeenCalledTimes(1);
-		expect(appMetricsObserver).toHaveBeenNthCalledWith(
-			1,
-			expect.objectContaining({
-				type: 'transport.request',
-				level: 'debug',
-				fields: expect.objectContaining({ status: 401, outcome: 'failed' }),
-			})
-		);
-		expect(appMetricsObserver).toHaveBeenNthCalledWith(
-			2,
-			expect.objectContaining({
-				type: 'transport.request',
-				level: 'error',
-				fields: expect.objectContaining({ status: 401 }),
-			})
-		);
-		fetch.mockRestore();
-	});
-
-	it('settles a 401 whose refresh yields no token as a warn row, not an error', async () => {
-		const credentials = { getLatest: jest.fn(() => ({ access_token: 'expired-token' })) };
-		const refreshAuth = jest.fn().mockResolvedValue(null);
-		const fetch = jest
-			.spyOn(globalThis, 'fetch')
-			.mockResolvedValue(new Response(null, { status: 401 }));
-		const { createAppSyncEngine, createRxdbSyncEngine, appMetricsObserver } = loadCreateAppEngine();
-		createAppSyncEngine({ ...BASE_OPTIONS, credentials, refreshAuth });
-		const fetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-
-		const response = await fetcher?.('https://store.example.test/wp-json/wcpos/v2/products');
-		await fetcher?.('https://store.example.test/wp-json/wcpos/v2/products');
-
-		// The refresh layer logs its own verdict (error when terminal); the request
-		// row records the failure without double-escalating.
-		expect(response?.status).toBe(401);
-		expect(fetch).toHaveBeenCalledTimes(2);
-		expect(appMetricsObserver).toHaveBeenCalledTimes(2);
-		for (const [event] of appMetricsObserver.mock.calls) {
-			expect(event).toEqual(
-				expect.objectContaining({
-					type: 'transport.request',
-					level: 'warn',
-					fields: expect.objectContaining({ status: 401, outcome: 'failed' }),
-				})
-			);
-			expect(event.fields).not.toHaveProperty('operationId');
-		}
-		fetch.mockRestore();
-	});
-
-	it('keeps the completion timestamp on a 401 attempt whose event is deferred', async () => {
-		jest.useFakeTimers().setSystemTime(1_000);
-		let accessToken = 'expired-token';
-		const credentials = { getLatest: jest.fn(() => ({ access_token: accessToken })) };
-		const refreshAuth = jest.fn(async () => {
-			jest.setSystemTime(5_000);
-			accessToken = 'refreshed-token';
-			return accessToken;
-		});
-		const fetch = jest
-			.spyOn(globalThis, 'fetch')
-			.mockResolvedValueOnce(new Response(null, { status: 401 }))
-			.mockResolvedValueOnce(new Response(null, { status: 200 }));
-
-		try {
-			const { createAppSyncEngine, createRxdbSyncEngine, appMetricsObserver } =
-				loadCreateAppEngine();
-			createAppSyncEngine({ ...BASE_OPTIONS, credentials, refreshAuth });
-			const fetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-
-			await fetcher?.('https://store.example.test/wp-json/wcpos/v2/products');
-
-			expect(appMetricsObserver).toHaveBeenNthCalledWith(
-				1,
-				expect.objectContaining({
-					type: 'transport.request',
-					at: 1_000,
-					fields: expect.objectContaining({ status: 401 }),
-				})
-			);
-		} finally {
-			fetch.mockRestore();
-			jest.useRealTimers();
-		}
-	});
-
-	it('classifies a 403 as error without ever refreshing (row-14 rule + #899 rubric)', async () => {
-		const credentials = { getLatest: jest.fn(() => ({ access_token: 'token' })) };
-		const refreshAuth = jest.fn();
-		const fetch = jest
-			.spyOn(globalThis, 'fetch')
-			.mockResolvedValueOnce(new Response(null, { status: 403 }));
-		const { createAppSyncEngine, createRxdbSyncEngine, appMetricsObserver } = loadCreateAppEngine();
-		createAppSyncEngine({ ...BASE_OPTIONS, credentials, refreshAuth });
-		const fetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-
-		const response = await fetcher?.('https://store.example.test/wp-json/wcpos/v2/products');
-
-		expect(response?.status).toBe(403);
-		expect(refreshAuth).not.toHaveBeenCalled();
-		expect(appMetricsObserver).toHaveBeenCalledWith(
-			expect.objectContaining({
-				type: 'transport.request',
-				level: 'error',
-				fields: expect.objectContaining({ status: 403 }),
-			})
-		);
-		fetch.mockRestore();
-	});
-
-	it('classifies the tick-probe 404 as a recovered debug row (designed fallback)', async () => {
-		const credentials = { getLatest: jest.fn(() => ({ access_token: 'token' })) };
-		const fetch = jest
-			.spyOn(globalThis, 'fetch')
-			.mockResolvedValueOnce(new Response(null, { status: 404 }));
-		const { createAppSyncEngine, createRxdbSyncEngine, appMetricsObserver } = loadCreateAppEngine();
-		createAppSyncEngine({ ...BASE_OPTIONS, credentials });
-		const fetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-
-		await fetcher?.('https://store.example.test/wp-json/wcpos/v2/changes/tick');
-
-		expect(appMetricsObserver).toHaveBeenCalledWith(
-			expect.objectContaining({
-				type: 'transport.request',
-				level: 'debug',
-				fields: expect.objectContaining({ status: 404, outcome: 'recovered' }),
-			})
-		);
-		fetch.mockRestore();
-	});
-
-	it('keeps a non-tick 404 at warn', async () => {
-		const credentials = { getLatest: jest.fn(() => ({ access_token: 'token' })) };
-		const fetch = jest
-			.spyOn(globalThis, 'fetch')
-			.mockResolvedValueOnce(new Response(null, { status: 404 }));
-		const { createAppSyncEngine, createRxdbSyncEngine, appMetricsObserver } = loadCreateAppEngine();
-		createAppSyncEngine({ ...BASE_OPTIONS, credentials });
-		const fetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-
-		await fetcher?.('https://store.example.test/wp-json/wcpos/v2/products');
-
-		expect(appMetricsObserver).toHaveBeenCalledWith(
-			expect.objectContaining({
-				type: 'transport.request',
-				level: 'warn',
-				fields: expect.objectContaining({ status: 404 }),
-			})
-		);
-		fetch.mockRestore();
-	});
-
-	it('retries with a peer-refreshed token instead of refreshing again on staggered 401s', async () => {
-		let accessToken = 'stale-token';
-		const credentials = { getLatest: jest.fn(() => ({ access_token: accessToken })) };
-		const refreshAuth = jest.fn(async () => {
-			accessToken = 'self-refreshed-token';
-			return accessToken;
-		});
-		const fetch = jest
-			.spyOn(globalThis, 'fetch')
-			.mockImplementationOnce(async () => {
-				// A concurrent request refreshes the JWT while this one is in flight.
-				accessToken = 'peer-refreshed-token';
-				return new Response(null, { status: 401 });
-			})
-			.mockResolvedValueOnce(new Response(null, { status: 200 }));
-		const { createAppSyncEngine, createRxdbSyncEngine } = loadCreateAppEngine();
-		createAppSyncEngine({ ...BASE_OPTIONS, credentials, refreshAuth });
-		const fetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-
-		const response = await fetcher?.('https://store.example.test/wp-json/wcpos/v2/products');
-
-		expect(response?.status).toBe(200);
-		expect(refreshAuth).not.toHaveBeenCalled();
-		expect(fetch).toHaveBeenCalledTimes(2);
-		const retryHeaders = fetch.mock.calls[1]?.[1]?.headers as Headers;
-		expect(retryHeaders.get('Authorization')).toBe('Bearer peer-refreshed-token');
-		fetch.mockRestore();
-	});
-
-	it('returns the original 401 when refresh fails without retrying', async () => {
-		const originalUnauthorized = new Response(null, { status: 401 });
-		const refreshAuth = jest.fn().mockResolvedValue(null);
-		const fetch = jest.spyOn(globalThis, 'fetch').mockResolvedValue(originalUnauthorized);
-		const { createAppSyncEngine, createRxdbSyncEngine, networkWarn } = loadCreateAppEngine();
-		createAppSyncEngine({ ...BASE_OPTIONS, refreshAuth });
-		const fetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-
-		const response = await fetcher?.('https://store.example.test/wp-json/wcpos/v2/products');
-
-		expect(response?.status).toBe(originalUnauthorized.status);
-		expect(refreshAuth).toHaveBeenCalledTimes(1);
-		expect(fetch).toHaveBeenCalledTimes(1);
-		// The unauthorized attempt is a failed unit of work and gets a terminal row.
-		expect(networkWarn).toHaveBeenCalledWith('transport.request', {
-			context: expect.objectContaining({ type: 'transport.request', status: 401 }),
-			terminal: expect.objectContaining({ operationType: 'sync.http', outcome: 'failed' }),
-		});
-		fetch.mockRestore();
-	});
-
-	it('does not refresh or loop when the retried request is still unauthorized', async () => {
-		const refreshAuth = jest.fn().mockResolvedValue('refreshed-token');
-		const fetch = jest
-			.spyOn(globalThis, 'fetch')
-			.mockResolvedValue(new Response(null, { status: 401 }));
-		const { createAppSyncEngine, createRxdbSyncEngine } = loadCreateAppEngine();
-		createAppSyncEngine({ ...BASE_OPTIONS, refreshAuth });
-		const fetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-
-		const response = await fetcher?.('https://store.example.test/wp-json/wcpos/v2/products');
-
-		expect(response?.status).toBe(401);
-		expect(refreshAuth).toHaveBeenCalledTimes(1);
-		expect(fetch).toHaveBeenCalledTimes(2);
-		fetch.mockRestore();
-	});
-
-	it('never refreshes a request to the refresh endpoint', async () => {
-		const refreshAuth = jest.fn().mockResolvedValue('refreshed-token');
-		const fetch = jest
-			.spyOn(globalThis, 'fetch')
-			.mockResolvedValue(new Response(null, { status: 401 }));
-		const { createAppSyncEngine, createRxdbSyncEngine } = loadCreateAppEngine();
-		createAppSyncEngine({ ...BASE_OPTIONS, refreshAuth });
-		const fetcher = createRxdbSyncEngine.mock.calls[0]?.[0].fetcher;
-
-		await fetcher?.('https://store.example.test/wp-json/wcpos/v2/auth/refresh');
-
-		expect(refreshAuth).not.toHaveBeenCalled();
-		expect(fetch).toHaveBeenCalledTimes(1);
 		fetch.mockRestore();
 	});
